Fixed FAB showing after hiding
The FAB was reappearing higher on the screen after being hidden (tested
on a Nexus 5 with API 22 & 23). This was due to the fact that
mYDisplayed was being assigned to a value lower than the actual Y
position of the FAB. Now, the hiding is done using a translation, which
I think is a bit more foolproof.